requirements.txt
Flask==3.0.3
requests==2.32.3
pymupdf==1.24.10
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
scipy==1.11.4